requirements.txt
# requirements.txt

annotated-types==0.7.0
appdirs==1.4.4
autograd==1.6.2
autoray==0.6.12
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
dill==0.3.8
filelock==3.15.4
fonttools==4.53.1
future==1.0.0
ibm-cloud-sdk-core==3.20.6
ibm-platform-services==0.55.3
idna==3.7
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.82
packaging==24.1
pandas==2.2.2
pbr==6.0.0
PennyLane==0.37.0
PennyLane-qiskit==0.37.0
PennyLane_Lightning==0.37.0
pillow==10.4.0
psutil==6.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pyparsing==3.1.2
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyspnego==0.11.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
qiskit==1.1.1
qiskit-aer==0.14.2
qiskit-ibm-provider==0.11.0
qiskit-ibm-runtime==0.27.0
requests==2.32.3
requests_ntlm==1.3.0
rustworkx==0.15.1
scipy==1.14.0
seaborn==0.13.2
semantic-version==2.10.0
six==1.16.0
sspilib==0.1.0
stevedore==5.2.0
symengine==0.11.0
sympy==1.12.1
toml==0.10.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.75
ultralytics-thop==2.0.0
urllib3==2.2.2
websocket-client==1.8.0
websockets==12.0
mss==6.1.0
pyautogui==0.9.53
roboflow==1.0.0
